Small perf improvements

+++ src/DotNetBoy.Emulator/Events/VBlankEventArgs.cs
@@ -4,5 +4,5 @@ namespace DotNetBoy.Emulator.Events;
 
 public class VBlankEventArgs : EventArgs
 {
-    public int[,] FrameBuffer { get; set; } = new int[ScreenDimensions.HEIGHT, ScreenDimensions.WIDTH];
+    public byte[,] FrameBuffer { get; set; } = new byte[ScreenDimensions.HEIGHT, ScreenDimensions.WIDTH];
 }
